+++ spec/integration/open_search/aws/sigv4_request_signer_spec.rb
@@ -13,17 +13,22 @@ require_relative '../../../spec_helper'
 require 'logger'
 require 'aws-sigv4'
 
-describe OpenSearch::Aws::Sigv4Client do
+describe OpenSearch::Aws::Sigv4RequestSigner do
   let(:client) do
     signer = Aws::Sigv4::Signer.new(service: 'es',
                                     region: 'us-west-2',
                                     access_key_id: 'key_id',
                                     secret_access_key: 'secret')
 
-    described_class.new({ host: OPENSEARCH_URL, logger: Logger.new($stdout) }, signer)
+    logger = Logger.new($stdout)
+    OpenSearch::Client.new({
+                             host: OPENSEARCH_URL,
+                             logger: logger,
+                             request_signer: described_class.new(signer)
+                           })
   end
 
-  it 'performs API actions without throwing any errors' do
+  it 'signs an API request without throwing any errors' do
     expect do
       # Index a document
       client.index(index: 'test-index', id: '1', body: { title: 'Test' })